+++ Team-B/src/test/java/org/springframework/samples/petclinic/pet/PetControllerTest.java
@@ -19,6 +19,7 @@ import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.util.Date;
 
+import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 @ExtendWith(SeleniumExtension.class)
@@ -67,4 +68,41 @@ public class PetControllerTest {
          utils.pause(5000);
 
      }
+
+    @Test //Amin This is the test that verifies if pet values can be updated
+    public void editPet_Test() throws InterruptedException {
+        String petName="Zizou";
+        String date="2000-01-02";
+        String type="dog";
+
+        utils.goToPage("owners?lastName=");
+
+        WebElement anOwner =driver.findElement(By.xpath("//table[@id='owners']/tbody/tr/td[1]/a"));
+        anOwner.click();
+
+        WebElement editPetLink = driver.findElement(By.linkText("Edit Pet"));
+        editPetLink.click();
+
+        WebElement petNameTxtBox = driver.findElement(By.id("name"));
+        petNameTxtBox.clear();
+        petNameTxtBox.sendKeys(petName);
+
+        WebElement birthDateTxtBox = driver.findElement(By.id("birthDate"));
+        birthDateTxtBox.clear();
+        birthDateTxtBox.sendKeys(date);
+
+        Select dropType = new Select(driver.findElement(By.name("type")));
+        dropType.selectByVisibleText(type);
+
+        WebElement updateButton = driver.findElement(By.xpath("//button[@type='submit']"));
+        updateButton.click();
+
+        WebElement newPetName =driver.findElement(By.xpath("//table[@class='table table-striped'][2]/tbody/tr/td/dl/dd[1]"));
+        WebElement newBirthDate =driver.findElement(By.xpath("//table[@class='table table-striped'][2]/tbody/tr/td/dl/dd[2]"));
+        WebElement newType =driver.findElement(By.xpath("//table[@class='table table-striped'][2]/tbody/tr/td/dl/dd[3]"));
+
+        assertEquals(newPetName.getText(),petName);
+        assertEquals(newBirthDate.getText(),date);
+        assertEquals(newType.getText(),type);
+    }
 }
